+++ packages/paymentChannel/src/index.ts
@@ -1,1 +1,5 @@
-console.log('stake');
+export type {
+  PaymentChannelState,
+  IPaymentChannel,
+} from './lib/paymentChannel.types';
+export { PaymentChannel } from './lib';
