--- packages/paymentChannel/src/lib/pws/index.ts
@@ -1,3 +1,0 @@
-export { PaymentChannelService } from "./service";
-export { IPaymentChannelService } from "./interface";
-export type { PaymentChannelState } from "./interface";
